Moved tests to filter field tests
