+++ graphene_django/tests/test_query.py
@@ -5,15 +5,12 @@ from django.db import models
 from django.utils.functional import SimpleLazyObject
 from py.test import raises
 
-from django_filters import FilterSet, NumberFilter
-
 import graphene
 from graphene.relay import Node
 
 from ..utils import DJANGO_FILTER_INSTALLED
 from ..compat import MissingType, JSONField
 from ..fields import DjangoConnectionField
-from ..filter.fields import DjangoFilterConnectionField
 from ..types import DjangoObjectType
 from .models import Article, Reporter
 
@@ -455,170 +452,3 @@ def test_should_query_node_multiple_filtering():
     result = schema.execute(query)
     assert not result.errors
     assert result.data == expected
-
-
-def test_should_query_filter_node_limit():
-    class ReporterFilter(FilterSet):
-        limit = NumberFilter(method='filter_limit')
-
-        def filter_limit(self, queryset, name, value):
-            return queryset[:value]
-
-        class Meta:
-            model = Reporter
-            fields = ['first_name', ]
-
-    class ReporterType(DjangoObjectType):
-
-        class Meta:
-            model = Reporter
-            interfaces = (Node, )
-
-    class ArticleType(DjangoObjectType):
-
-        class Meta:
-            model = Article
-            interfaces = (Node, )
-            filter_fields = ('lang', )
-
-    class Query(graphene.ObjectType):
-        all_reporters = DjangoFilterConnectionField(
-            ReporterType,
-            filterset_class=ReporterFilter
-        )
-
-        def resolve_all_reporters(self, args, context, info):
-            return Reporter.objects.order_by('a_choice')
-
-    Reporter.objects.create(
-        first_name='Bob',
-        last_name='Doe',
-        email='bobdoe@example.com',
-        a_choice=2
-    )
-    r = Reporter.objects.create(
-        first_name='John',
-        last_name='Doe',
-        email='johndoe@example.com',
-        a_choice=1
-    )
-
-    Article.objects.create(
-        headline='Article Node 1',
-        pub_date=datetime.date.today(),
-        reporter=r,
-        editor=r,
-        lang='es'
-    )
-    Article.objects.create(
-        headline='Article Node 2',
-        pub_date=datetime.date.today(),
-        reporter=r,
-        editor=r,
-        lang='en'
-    )
-
-    schema = graphene.Schema(query=Query)
-    query = '''
-        query NodeFilteringQuery {
-            allReporters(limit: 1) {
-                edges {
-                    node {
-                        id
-                        firstName
-                        articles(lang: "es") {
-                            edges {
-                                node {
-                                    id
-                                    lang
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    '''
-
-    expected = {
-        'allReporters': {
-            'edges': [{
-                'node': {
-                    'id': 'UmVwb3J0ZXJUeXBlOjI=',
-                    'firstName': 'John',
-                    'articles': {
-                        'edges': [{
-                            'node': {
-                                'id': 'QXJ0aWNsZVR5cGU6MQ==',
-                                'lang': 'ES'
-                            }
-                        }]
-                    }
-                }
-            }]
-        }
-    }
-
-    result = schema.execute(query)
-    assert not result.errors
-    assert result.data == expected
-
-
-def test_should_query_filter_node_double_limit_raises():
-    class ReporterFilter(FilterSet):
-        limit = NumberFilter(method='filter_limit')
-
-        def filter_limit(self, queryset, name, value):
-            return queryset[:value]
-
-        class Meta:
-            model = Reporter
-            fields = ['first_name', ]
-
-    class ReporterType(DjangoObjectType):
-
-        class Meta:
-            model = Reporter
-            interfaces = (Node, )
-
-    class Query(graphene.ObjectType):
-        all_reporters = DjangoFilterConnectionField(
-            ReporterType,
-            filterset_class=ReporterFilter
-        )
-
-        def resolve_all_reporters(self, args, context, info):
-            return Reporter.objects.order_by('a_choice')[:2]
-
-    Reporter.objects.create(
-        first_name='Bob',
-        last_name='Doe',
-        email='bobdoe@example.com',
-        a_choice=2
-    )
-    r = Reporter.objects.create(
-        first_name='John',
-        last_name='Doe',
-        email='johndoe@example.com',
-        a_choice=1
-    )
-
-    schema = graphene.Schema(query=Query)
-    query = '''
-        query NodeFilteringQuery {
-            allReporters(limit: 1) {
-                edges {
-                    node {
-                        id
-                        firstName
-                    }
-                }
-            }
-        }
-    '''
-
-    result = schema.execute(query)
-    assert len(result.errors) == 1
-    assert str(result.errors[0]) == (
-        'Received two sliced querysets (high mark) in the connection, please slice only in one.'
-    )
